requirements.txt
# KonveyN2AI Dependencies - PINNED VERSIONS FOR CONSISTENCY
# FastAPI 0.116.1 resolves TestClient compatibility with httpx 0.28.1
#
# CRITICAL: All versions are pinned (==) to prevent CI/CD conflicts
# Lessons learned from Task 1 protocol violations:
# - Unpinned versions caused version conflicts and pipeline failures
# - Multiple fix commits were required due to dependency mismatches
# - "Test locally first" protocol prevents these issues
#
# Reference: .memory/tasks/task-1-project-setup.md - Protocol Violations section
# Exception context: Git consolidation exception for hackathon time constraints
# Date: 2025-07-26

# Core Google Cloud AI Platform dependencies
google-cloud-aiplatform==1.42.1
google-cloud-artifact-registry==1.16.1
google-genai==1.27.0

# BigQuery Vector Backend dependencies (NEW)
google-cloud-bigquery==3.25.0
google-generativeai==0.7.2
scikit-learn==1.4.0
joblib==1.3.2

# Web framework and API
streamlit==1.30.0
fastapi==0.116.1
uvicorn[standard]==0.25.0

# Data processing and numerical computing
numpy==1.26.3
pandas==2.2.0

# Environment and configuration management
python-dotenv==1.0.0
pydantic==2.6.1

# Development and code quality tools (pinned versions for consistency)
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.0.0
pytest-mock==3.12.0

black==25.1.0
ruff==0.1.15
pre-commit==3.6.0
mypy==1.8.0
isort==5.13.2
bandit==1.7.6

# Additional utilities for hackathon development
requests==2.31.0
aiohttp==3.9.3
tenacity==8.2.3
httpx==0.28.1

# M1 Ingestion dependencies
pyyaml==6.0.1
kubernetes==29.0.0
kr8s==0.10.0
libcst==1.1.0
ulid-py==1.1.0
python-cobol==0.1.4
google-cloud-run==0.11.0
